doc_search/requirements.txt
python-dotenv
elasticsearch==8.16.0
sentence-transformers
google-generativeai
typer[all]
